+++ Programming/lab5-mitrian/common/src/main/java/com/mitrian/lab/common/data/initializer/IdCollection.java
@@ -6,12 +6,11 @@ public class IdCollection {
 
     public static HashSet<Long> idCollection = new HashSet<>();
 
-    public static long createWorkerId(){
-        long id;
+    public static Integer createWorkerId(){
+        Integer id;
         do{
-            id = (System.currentTimeMillis() & 0xffff);
+            id = Math.toIntExact((System.currentTimeMillis() & 0xffff));
         } while (idCollection.contains(id));
         return id;
     }
-
 }
